+++ inst/htmlwidgets/scatterD3-labels.js
@@ -1,6 +1,4 @@
 
-
-
 function labels_create(chart) {
 
     if (!chart.settings().has_labels) return;
@@ -11,7 +9,7 @@ function labels_create(chart) {
     
     labels.enter()
         .append("text")
-	.call(label_init)
+	    .call(label_init)
         .call(label_formatting, chart)
         .call(drag_behavior(chart));
 
@@ -26,13 +24,14 @@ function labels_update(chart) {
     function endall(transition, callback) {
         if (typeof callback !== "function") throw new Error("Wrong callback in endall");
         if (transition.size() === 0) { callback() }
-        var n = 0;
+        let n = 0;
         transition
-            .each(function() { ++n; })
-            .on("end", function() { if (!--n) callback.apply(this, arguments); });
+            .each(() => (++n))
+            .on("end", () => { if (!--n) callback.apply(this, arguments); });
       }
 
     if (!chart.settings().has_labels) return;
+    if (chart.settings().positions_changed) labels_placement(chart); 
 
     var labels = chart.svg().select(".chart-body")
         .selectAll(".point-label")
@@ -42,9 +41,11 @@ function labels_update(chart) {
         .call(label_init)
         .call(drag_behavior(chart))
         .merge(labels)
-        .transition().duration(1100)
+        .transition().duration(1000)
         .call(label_formatting, chart)
-        .call(endall, function() { labels_placement(chart); });
+        // .call(endall, () => {
+        //     if (chart.settings().positions_changed) labels_placement(chart); 
+        // });
 
     labels.exit()
         .each(function(d) {
@@ -87,10 +88,14 @@ function default_label_dy(d, chart) {
 function get_label_dx(d, i, chart) {
     // Manually defined
     if (d.lab_dx !== undefined) return (d.lab_dx);
+    // From chart.positions()
+    const position = chart.positions().filter(p => key(p) == key(d))
+    if (position.length > 0) {
+        return position[0].lab_dx;
+    }
     // From labels_positions argument
     if (chart.settings().labels_positions && chart.settings().labels_positions != "auto") {
-        var dx = chart.scales().x(chart.settings().labels_positions[i].lab_x) - chart.scales().x(d.x);
-        return dx;
+        return chart.scales().x(chart.settings().labels_positions[i].lab_x) - chart.scales().x(d.x);
     }
     // Default
     return (0);
@@ -100,6 +105,11 @@ function get_label_dx(d, i, chart) {
 function get_label_dy(d, i, chart) {
     // Manually defined
     if (d.lab_dy !== undefined) return (d.lab_dy);
+    // From chart.positions()
+    const position = chart.positions().filter(p => key(p) == key(d))
+    if (position.length > 0) {
+        return position[0].lab_dy;
+    }
     // From labels_positions argument
     if (chart.settings().labels_positions && chart.settings().labels_positions != "auto") {
         var dy = chart.scales().y(chart.settings().labels_positions[i].lab_y) - chart.scales().y(d.y);
@@ -113,94 +123,19 @@ function get_label_dy(d, i, chart) {
 function label_formatting(selection, chart) {
 
     selection
-	.filter(function(d) { return d.lab !== "" && d.lab !== null; })
-        .text(function (d) { return (d.lab); })
+	    .filter(d => (d.lab !== "" && d.lab !== null))
+        .text(d => (d.lab))
         .style("font-size", chart.settings().labels_size + "px")
-        .attr("class", function (d, i) {
-            return "point-label color color-c" + css_clean(d.col_var) + " symbol symbol-c" + css_clean(d.symbol_var);
-        })
-        .attr("transform", function (d) { return translation(d, chart.scales()); })
-        .style("fill", function (d) { return chart.scales().color(d.col_var); })
-        .attr("dx", function(d, i) { return get_label_dx(d, i, chart) + "px"; })
-        .attr("dy", function(d, i) { return get_label_dy(d, i, chart) + "px"; })
-        .each(function(d, i) {
-            var label = d3v5.select(this);
-            var dx = get_label_dx(d, i, chart);
-            var dy = get_label_dy(d, i, chart);
-            label.call(label_line_formatting, d, dx, dy, chart);
-        })
-}
-
-
-// Compute end of label line coordinates and distance with point
-function label_line_coordinates(label, x_orig, y_orig, x, y) {
-    
-    var label_bb = label.node().getBBox();
-    var bb = {left: x - label_bb.width / 2,
-              right: x + label_bb.width / 2,
-              top: y - 3 * label_bb.height / 4,
-              bottom: y + label_bb.height / 4,
-              middle: y - label_bb.height / 4};
-    var coord = {};
-
-    if (bb.left > x_orig + 4) {
-        coord.x = bb.left - 5;
-        coord.y = bb.middle;
-    } else if (bb.right < x_orig - 4) {
-        coord.x = bb.right + 5;
-        coord.y = bb.middle;
-    } else {
-        coord.x = x;
-        if (y > y_orig) {
-            coord.y = bb.top;
-        } else {
-            coord.y = bb.bottom;
-        }
-    }
-
-    coord.dist = Math.sqrt(Math.pow(coord.x - x_orig, 2) + Math.pow(coord.y - y_orig, 2));
-
-    // No line if label is just around point
-    if (bb.bottom  >= y_orig - 10 && bb.top <= y_orig + 10 &&
-        bb.left <= x_orig + 4 && bb.right >= x_orig - 4) {
-        coord.dist = 0;
-    }
-
-    return(coord);
+        .attr("class", (d, i) => 
+            (`point-label color color-c${css_clean(d.col_var)} symbol symbol-c${css_clean(d.symbol_var)}`)
+        )
+        .attr("transform", d => ( translation(d, chart.scales()) ))
+        .style("fill",   d => ( chart.scales().color(d.col_var) ))
+        .attr("dx", (d, i) => ( get_label_dx(d, i, chart) + "px" ))
+        .attr("dy", (d, i) => ( get_label_dy(d, i, chart) + "px" ))
 
 }
 
-// Format line between point and label
-function label_line_formatting(selection, d, dx, dy, chart) {
-
-    var x = chart.scales().x(d.x);
-    var y = chart.scales().y(d.y);
-    var coord = label_line_coordinates(selection, x, y, x + dx, y + dy);
-    var x2 = coord.x - x;
-    var y2 = coord.y - y;
-    var line = chart.svg().select(".label-line-" + css_clean(key(d)));
-    // Force negative gap for labels below arrows
-    var gap0 = -Math.abs(default_label_dy(d, chart));
-
-    if (coord.dist > 15 && line.empty()) {
-        line = chart.svg().select(".chart-body")
-            .append("svg:line")
-            .lower()
-            .datum(d)
-            .attr("transform", translation(d, chart.scales()))
-            .attr("class", function (d, i) {
-                return "point-label-line label-line-" + css_clean(key(d)) + " color color-c" + css_clean(d.col_var) + " symbol symbol-c" + css_clean(d.symbol_var);
-            })
-    }
-    if (coord.dist > 15) {
-        line.attr("x1", - x2 * gap0 / coord.dist).attr("x2", x2)
-            .attr("y1", - y2 * gap0 / coord.dist).attr("y2", y2)
-            .style("stroke", chart.scales().color(d.col_var));
-    }
-    if (coord.dist <= 15 && !line.empty()) {
-        line.remove();
-    }
-}
 
 
 // Compute automatic label placement
@@ -215,7 +150,7 @@ function labels_placement(chart) {
 
     var labels = chart.svg().selectAll(".point-label");
 
-    labels = labels.filter(function(d) { return d.lab !== "" && d.lab !== null;});
+    labels = labels.filter(d => (d.lab !== "" && d.lab !== null));
 
     labels.each(function (d) {
         var bb = this.getBBox();
@@ -239,15 +174,20 @@ function labels_placement(chart) {
         .height(chart.dims().height)
         .start(nsweeps);
 
+    let positions = chart.positions();        
     labels.data().forEach(function (d, i) {
         d.lab_dx = label_array[i].x - chart.scales().x(d.x);
         d.lab_dy = label_array[i].y - chart.scales().y(d.y);
+        positions = positions.filter(p => key(p) != key(d))
+        positions.push({lab_dx: d.lab_dx, lab_dy: d.lab_dy, key_var: d.key_var})
     })
+    chart.positions(positions);
 
     labels
         .transition().duration(1000)
-        .call(label_formatting, chart);
+        .call(label_formatting, chart)
 
+    labels.call(label_line_display, chart)
 }
 
 
@@ -265,31 +205,36 @@ function drag_behavior(chart) {
         })
         .on('start', function (d) {
             if (!d3v5.event.sourceEvent.shiftKey) {
-                dragging = true;
+                chart.dragging(true);
                 var label = d3v5.select(this);
                 label.style('fill', '#000');
                 var dx = d3v5.event.x - scales.x(d.x);
                 var dy = d3v5.event.y - scales.y(d.y);
-                label.call(label_line_formatting, d, dx, dy, chart);
+                label.call(label_line_display, chart);
             }
         })
-        .on('drag', function(d) {
-            if (dragging) {
+        .on('drag', function(d, i) {
+            if (chart.dragging()) {
                 var label = d3v5.select(this);
                 var dx = d3v5.event.x - scales.x(d.x);
                 var dy = d3v5.event.y - scales.y(d.y);
                 label.attr('dx', dx + "px")
                      .attr('dy', dy + "px");
-                label.call(label_line_formatting, d, dx, dy, chart);
                 d.lab_dx = dx;
                 d.lab_dy = dy;
+                label.call(label_line_display, chart);
+                let positions = chart.positions();
+                positions = positions.filter(p => key(p) != key(d))
+                positions.push({lab_dx: d.lab_dx, lab_dy: d.lab_dy, key_var: d.key_var})        
+                chart.positions(positions)
             }
         })
         .on('end', function(d) {
-            if (dragging) {
-                d3v5.select(this).style('fill', chart.scales().color(d.col_var));
-                d3v5.select(this).call(label_line_formatting, d, d.lab_dx, d.lab_dy, chart);
-                dragging = false;
+            if (chart.dragging()) {
+                var label = d3v5.select(this);
+                label.style('fill', chart.scales().color(d.col_var));
+                label.call(label_line_display, chart);
+                chart.dragging(false);
             }
         });